fixing bad conflicts and imports
